Monty

svn path=/trunk/ogg/; revision=989

+++ include/ogg/ogg.h
@@ -35,10 +35,10 @@ typedef struct {
   long    body_returned;         /* elements of fill returned */
 
 
-  int     *lacing_vals;    /* The values that will go to the segment table */
-  ogg_int64_t *granule_vals;      /* granulepos values for headers. Not compact
-                             this way, but it is simple coupled to the
-                             lacing fifo */
+  int     *lacing_vals;      /* The values that will go to the segment table */
+  ogg_int64_t *granule_vals; /* granulepos values for headers. Not compact
+				this way, but it is simple coupled to the
+				lacing fifo */
   long    lacing_storage;
   long    lacing_fill;
   long    lacing_packet;
@@ -51,8 +51,8 @@ typedef struct {
                              logical bitstream */
   int     b_o_s;          /* set after we've written the initial page
                              of a logical bitstream */
-  long     serialno;
-  int      pageno;
+  long    serialno;
+  long    pageno;
   ogg_int64_t  packetno;      /* sequence number for decode; the framing
                              knows where there's a hole in the data,
                              but we need coupling so that the codec
@@ -72,12 +72,12 @@ typedef struct {
   long  e_o_s;
 
   ogg_int64_t  granulepos;
-  ogg_int64_t  packetno;       /* sequence number for decode; the framing
-                             knows where there's a hole in the data,
-                             but we need coupling so that the codec
-                             (which is in a seperate abstraction
-                             layer) also knows about the gap */
-
+  
+  ogg_int64_t  packetno;     /* sequence number for decode; the framing
+				knows where there's a hole in the data,
+				but we need coupling so that the codec
+				(which is in a seperate abstraction
+				layer) also knows about the gap */
 } ogg_packet;
 
 typedef struct {
@@ -144,8 +144,8 @@ extern int      ogg_page_bos(ogg_page *og);
 extern int      ogg_page_eos(ogg_page *og);
 extern ogg_int64_t  ogg_page_granulepos(ogg_page *og);
 extern int      ogg_page_serialno(ogg_page *og);
-extern int      ogg_page_pageno(ogg_page *og);
-  extern int      ogg_page_packets(ogg_page *og);
+extern long     ogg_page_pageno(ogg_page *og);
+extern int      ogg_page_packets(ogg_page *og);
 
 
 #ifdef __cplusplus
